Draft 0.1 Atmospherics Department
This draft only has the main atmospherics department and new "canister" storage. I'll likely change it / move it, but for now, this is what I've gotten done and I'm clearly not as talented with mapping (and atmospherics) as others.

If nobody else adds to this, I will get it in soon.
If anyone else is more confident than me, I am more than willing to let you continue off of my work.