Atmos Touch Ups
This should mostly just be the change I made to the security department as a "proof of concept". Admittedly, the map is so congested and jointed together than I don't know how to add what I was asked to add. I was able to sorta do it with the security department? But I'll leave it up to Cadyn and the rest to decide whether or not this is what we want.

MIND YOU, I HAVE YET TO TEST RUN ANYTHING. I will be doing that after this commit.